# contracts/openapi/print-platform.yaml
openapi: 3.1.0
info:
  title: Print Platform API
  version: 0.3.0
  description: |
    Print Platform contract for print tasks, template lifecycle, template data configuration,
    template test runs, AI-assisted layout, and multi-database source connections.

    Template data/testing endpoints are implemented. Test runs are isolated from formal print tasks,
    and SQL execution uses backend-only read-only data source connections.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/print-previews:
    post:
      tags: [Business Integration]
      operationId: createPrintPreview
      description: |
        Resolves the currently published template version and runtime business data for an embedded
        business-system preview. This endpoint does not create a PrintTask and does not talk to the
        browser extension or local Print Agent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePrintPreviewRequest' }
      responses:
        '200':
          description: Published template design and resolved runtime render data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrintPreview' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/print-tasks:
    get:
      tags: [Print Tasks]
      operationId: listPrintTasks
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PrintTaskStatus' } }
        - { name: templateCode, in: query, schema: { type: string } }
        - { name: printerId, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 200 } }
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PrintTask' }
    post:
      tags: [Print Tasks]
      operationId: createPrintTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePrintTaskRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrintTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/system-logs:
    get:
      tags: [Runtime]
      operationId: listSystemLogs
      parameters:
        - { name: level, in: query, schema: { type: string, enum: [DEBUG, INFO, WARN, ERROR, FATAL] } }
        - { name: module, in: query, schema: { type: string } }
        - { name: eventType, in: query, schema: { type: string } }
        - { name: taskId, in: query, schema: { type: string } }
        - { name: attemptId, in: query, schema: { type: integer } }
        - { name: agentId, in: query, schema: { type: string } }
        - { name: printerId, in: query, schema: { type: string } }
        - { name: spoolerJobId, in: query, schema: { type: integer } }
        - { name: resolved, in: query, schema: { type: boolean } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Filtered system log page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemLogPage' }
...
  /api/system-logs/batch:
    post:
      tags: [Runtime]
      operationId: uploadSystemLogs
      description: Agent batch upload endpoint. Maximum 200 entries per request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 200
              items: { $ref: '#/components/schemas/SystemLogBatchItemRequest' }
      responses:
        '200':
          description: Accepted batch count
...
  /api/agents/heartbeat:
    post:
      tags: [Runtime]
      operationId: agentHeartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentHeartbeatRequest' }
      responses:
        '200':
          description: Heartbeat accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentHeartbeatResponse' }
...
  /api/alerts:
    get:
      tags: [Runtime]
      operationId: listAlerts
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [OPEN, ACKED, RESOLVED] } }
        - { name: severity, in: query, schema: { type: string } }
        - { name: ruleCode, in: query, schema: { type: string } }
        - { name: resourceType, in: query, schema: { type: string } }
        - { name: agentId, in: query, schema: { type: string } }
        - { name: printerId, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Alert page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertPage' }
...
  /api/templates:
    get:
      tags: [Templates]
      operationId: listTemplates
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PrintTemplate' }
    post:
      tags: [Templates]
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTemplateRequest' }
      responses:
        '201':
          description: Created template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrintTemplate' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Print Platform API Key sent as Authorization: Bearer <key>."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Print-Api-Key
      description: Alternative header for Print Platform API Key authentication.
...
  responses:
...
    BadRequest:
      description: Invalid request or static configuration
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    Conflict:
      description: Concurrent modification or invalid state transition
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable machine-readable platform error code, for example TEMPLATE_NOT_FOUND, TEMPLATE_STATE_CONFLICT, DATA_CONFIG_CORRUPTED, SECRET_REF_INVALID, or SECRET_NOT_RESOLVED.
        queryId: { type: [string, 'null'] }
        testRunId: { type: [string, 'null'] }
      additionalProperties: true
...
    CreatePrintPreviewRequest:
      type: object
      required: [templateCode]
      properties:
        templateCode: { type: string, minLength: 1 }
        params:
          type: object
          additionalProperties: true
          description: Runtime SQL parameters for a published SQL-mode template.
        inputData:
          oneOf:
            - { $ref: '#/components/schemas/TemplateSampleData' }
            - { type: 'null' }
          description: Required for published JSON-mode templates. SQL mode rejects this field.
    PrintPreview:
      type: object
      required: [templateCode, templateVersion, templateName, documentType, design, renderData, generatedAt]
      properties:
        templateCode: { type: string }
        templateVersion: { type: integer, minimum: 1 }
        templateName: { type: string }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        design: { $ref: '#/components/schemas/TemplateDesign' }
        renderData:
          type: object
          additionalProperties: true
        generatedAt: { type: string, format: date-time }
    CreatePrintTaskRequest:
      type: object
      required: [templateCode, businessKey, printerId, copies]
      properties:
        templateCode: { type: string, minLength: 1 }
        businessKey: { type: string, minLength: 1 }
        printerId: { type: [string, 'null'] }
        copies: { type: integer, minimum: 1 }
        params:
          type: object
          additionalProperties: true
          description: Runtime SQL parameters for a published SQL-mode template.
        inputData:
          oneOf:
            - { $ref: '#/components/schemas/TemplateSampleData' }
            - { type: 'null' }
          description: Required for a formal JSON-mode print task. The backend never falls back to template sampleData for formal printing. SQL mode rejects this field.
...
    PrintTaskStatus:
      type: string
      enum: [CREATED, QUEUED, PRINTING, SUCCESS, FAILED, RETRYING, WAITING_AGENT, CANCELLED]
    PrintTask:
      type: object
      required: [id, templateCode, businessKey, snapshotId, copies, status, attempts, createdAt, updatedAt]
      properties:
        id: { type: string }
        templateCode: { type: string }
        templateVersion: { type: [integer, 'null'], minimum: 1 }
        businessKey: { type: string }
        snapshotId: { type: string }
        printerId: { type: [string, 'null'] }
        copies: { type: integer }
        status: { $ref: '#/components/schemas/PrintTaskStatus' }
        attempts: { type: integer }
        message: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    SystemLogEntry:
      type: object
      required: [id, level, module, eventType, message, resolved, createdAt, updatedAt]
      properties:
        id: { type: integer }
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR, FATAL] }
        module: { type: string }
        eventType: { type: string }
        message: { type: string }
        stackTrace: { type: [string, 'null'] }
        requestId: { type: [string, 'null'] }
        taskId: { type: [string, 'null'] }
        attemptId: { type: [integer, 'null'] }
        agentId: { type: [string, 'null'] }
        agentInstanceId: { type: [string, 'null'] }
        printerId: { type: [string, 'null'] }
        spoolerJobId: { type: [integer, 'null'] }
        hostName: { type: [string, 'null'] }
        ipAddress: { type: [string, 'null'] }
        resolved: { type: boolean }
        resolvedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SystemLogPage:
      type: object
      required: [total, page, size, items]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 200 }
        items:
          type: array
          items: { $ref: '#/components/schemas/SystemLogEntry' }
    SystemLogBatchItemRequest:
      type: object
      required: [level, module, eventType, message]
      properties:
        level: { type: string, enum: [DEBUG, INFO, WARN, ERROR, FATAL] }
        module: { type: string, maxLength: 64 }
        eventType: { type: string, maxLength: 128 }
        message: { type: string, maxLength: 8000 }
        stackTrace: { type: [string, 'null'], maxLength: 32000 }
        requestId: { type: [string, 'null'], maxLength: 64 }
        taskId: { type: [string, 'null'], maxLength: 64 }
        attemptId: { type: [integer, 'null'] }
        agentId: { type: [string, 'null'], maxLength: 128 }
        agentInstanceId: { type: [string, 'null'], maxLength: 128 }
        printerId: { type: [string, 'null'], maxLength: 128 }
        spoolerJobId: { type: [integer, 'null'] }
        hostName: { type: [string, 'null'], maxLength: 128 }
        ipAddress: { type: [string, 'null'], maxLength: 64 }
        createdAt: { type: [string, 'null'], format: date-time }
...
    AgentHeartbeatRequest:
      type: object
      required: [agentId, instanceId, hostName, osName, agentVersion]
      properties:
        agentId: { type: string, maxLength: 128 }
        instanceId: { type: string, maxLength: 128 }
        hostName: { type: string, maxLength: 128 }
        ipAddress: { type: [string, 'null'], maxLength: 64 }
        osName: { type: string, maxLength: 128 }
        agentVersion: { type: string, maxLength: 64 }
        cpuUsage: { type: [number, 'null'], minimum: 0, maximum: 100 }
        memoryUsage: { type: [number, 'null'], minimum: 0, maximum: 100 }
        activeJobs: { type: [integer, 'null'], minimum: 0 }
        queuedJobs: { type: [integer, 'null'], minimum: 0 }
        printerCount: { type: [integer, 'null'], minimum: 0 }
        spoolerStatus: { type: [string, 'null'], maxLength: 32 }
        printers:
          type: [array, 'null']
          maxItems: 256
          description: |
            Authoritative printer snapshot for this heartbeat. Omit this field for legacy agents that only report printerCount.
            An explicit empty array means the node currently has no discovered printers.
          items: { $ref: '#/components/schemas/AgentPrinterReport' }
    AgentPrinterReport:
      type: object
      required: [printerId, name, type, status]
      properties:
        printerId: { type: string, minLength: 1, maxLength: 256 }
        name: { type: string, minLength: 1, maxLength: 256 }
        type: { type: string, minLength: 1, maxLength: 64 }
        status: { type: string, minLength: 1, maxLength: 64 }
        defaultPrinter: { type: [boolean, 'null'] }
        driverName: { type: [string, 'null'], maxLength: 256 }
        portName: { type: [string, 'null'], maxLength: 256 }
        shared: { type: [boolean, 'null'] }
        shareName: { type: [string, 'null'], maxLength: 256 }
        location: { type: [string, 'null'], maxLength: 256 }
        comment: { type: [string, 'null'], maxLength: 512 }
        paperSizes:
          type: [array, 'null']
          maxItems: 64
          items: { type: string, minLength: 1, maxLength: 128 }
...
    AgentHeartbeatResponse:
      type: object
      required: [status, nextHeartbeatSeconds, serverTime]
      properties:
        status: { type: string }
        nextHeartbeatSeconds: { type: integer }
        serverTime: { type: string, format: date-time }
...
    AlertEvent:
      type: object
      required: [id, ruleCode, severity, status, resourceType, resourceId, message, firstSeenAt, lastSeenAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        ruleCode: { type: string }
        severity: { type: string }
        status: { type: string, enum: [OPEN, ACKED, RESOLVED] }
        resourceType: { type: string }
        resourceId: { type: string }
        message: { type: string }
        taskId: { type: [string, 'null'] }
        attemptId: { type: [integer, 'null'] }
        agentId: { type: [string, 'null'] }
        printerId: { type: [string, 'null'] }
        spoolerJobId: { type: [integer, 'null'] }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        ackedAt: { type: [string, 'null'], format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AlertPage:
      type: object
      required: [total, page, size, items]
      properties:
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/AlertEvent' }
...
    TemplateStatus:
      type: string
      enum: [DRAFT, TESTING, REVIEWING, PUBLISHED, DISABLED]
    DocumentType:
      type: string
      enum: [FORM, INVOICE, RECEIPT, EXPENSE_LIST, POS_RECEIPT, LABEL, REPORT]
    TemplateDataMode:
      type: string
      enum: [JSON, SQL]
    TemplateQueryResultType:
      type: string
      enum: [OBJECT, ARRAY, SCALAR]
    TemplateQueryParamType:
      type: string
      enum: [STRING, INTEGER, DECIMAL, BOOLEAN, DATE, DATETIME]
    TemplateQueryParam:
      type: object
      required: [name, type, required]
      properties:
        name:
          type: string
          pattern: '^[A-Za-z_][A-Za-z0-9_]{0,63}$'
        type: { $ref: '#/components/schemas/TemplateQueryParamType' }
        required: { type: boolean }
        defaultValue:
          description: Scalar default value converted according to param type.
          oneOf:
            - { type: string }
            - { type: integer }
            - { type: number }
            - { type: boolean }
            - { type: 'null' }
    TemplateQueryConfig:
      type: object
      required: [queryId, name, connectionId, sql, params, resultKey, resultType, order, enabled]
      properties:
        queryId:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,63}$'
        name: { type: string, minLength: 1 }
        connectionId: { type: string, minLength: 1 }
        sql:
          type: string
          minLength: 1
          description: One read-only SELECT or WITH...SELECT statement using named parameters.
        params:
          type: array
          items: { $ref: '#/components/schemas/TemplateQueryParam' }
        resultKey:
          type: string
          pattern: '^[A-Za-z_][A-Za-z0-9_]{0,63}$'
          not:
            enum: [__proto__, prototype, constructor]
        resultType: { $ref: '#/components/schemas/TemplateQueryResultType' }
        order: { type: integer }
        enabled: { type: boolean }
        timeoutMs:
          type: integer
          minimum: 100
          maximum: 30000
        maxRows:
          type: integer
          minimum: 1
          maximum: 10000
    TemplateDataConfig:
      type: object
      required: [schemaVersion, mode, queries]
      properties:
        schemaVersion:
          type: integer
          const: 1
        mode: { $ref: '#/components/schemas/TemplateDataMode' }
        queries:
          type: array
          items: { $ref: '#/components/schemas/TemplateQueryConfig' }
      examples:
        - schemaVersion: 1
          mode: JSON
          queries: []
    TemplateDesign:
      type: object
      additionalProperties: true
      description: Existing visual template design JSON.
    TemplateSampleData:
      type: object
      additionalProperties: true
      description: Stable sample data contract used for design-time preview and binding validation.
    CreateTemplateRequest:
      type: object
      required: [code, name, documentType]
      properties:
        code: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        documentType: { $ref: '#/components/schemas/DocumentType' }
...
    PrintTemplate:
      type: object
      required:
        [id, code, name, documentType, status, draftRevision, design, sampleData, dataConfig, createdAt, updatedAt]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        status: { $ref: '#/components/schemas/TemplateStatus' }
        draftRevision: { type: integer, minimum: 1 }
        publishedVersion: { type: [integer, 'null'], minimum: 1 }
        design: { $ref: '#/components/schemas/TemplateDesign' }
        sampleData: { $ref: '#/components/schemas/TemplateSampleData' }
        dataConfig: { $ref: '#/components/schemas/TemplateDataConfig' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...